Extracts response parser to its module
